apollo_l1_gas_price: Chainlink retdata decoding and rate arithmetic (#14989)

+++ crates/apollo_l1_gas_price_types/src/errors.rs
@@ -50,7 +50,7 @@ pub enum ExchangeRateOracleClientError {
     #[error("Missing or invalid field: {0}. Body: {1}")]
     MissingFieldError(String, String),
     #[error("Invalid decimals value: expected {0}, got {1}")]
-    InvalidDecimalsError(u64, u64),
+    InvalidDecimalsError(u32, u32),
     #[error("Query not yet resolved: timestamp={0}")]
     QueryNotReadyError(u64),
     #[error("All URLs in the list failed for timestamp {0}, starting with index {1}")]
